Consolidate prompts
