+++ src/lib.rs
@@ -1,5 +1,4 @@
 pub mod polkadot_sdk_releases;
-pub mod prdoc_analysis_prompts;
 pub mod prompts;
 pub mod public_endpoints;
 pub mod resources;
